requirements-dev.txt
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pytest-cov==4.1.0
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
aiosqlite==0.19.0
sqlalchemy==2.0.30
fastapi==0.110.1
uvicorn==0.29.0
pydantic-settings==2.2.1
passlib==1.7.4
asyncpg==0.29.0
alembic==1.13.1
langchain
langgraph
openai
anthropic
bcrypt
python-jose[cryptography]
python-jwt==4.1.0
psycopg2-binary==2.9.9